[Bronze V] Title: A+B - 9, Time: 0 ms, Memory: 2020 KB, Score: 75 point -BaekjoonHub

+++ 백준/Bronze/15740. A＋B － 9/A＋B － 9.cc
@@ -3,8 +3,8 @@
 int main () {
     std::cin.tie(NULL);
     std::ios::sync_with_stdio(false);
-    int a, b;
-    int sum;
+    long long int a, b;
+    long long int sum;
 
     std::cin >> a >> b;
